+++ NEWS.md
@@ -1,7 +1,7 @@
 # Version 0.0.4-9000 (2025-10-07)
 
  * Add support for futurizing calls nested in, one or more layers, of
-   `{ ... }`, `( ... )`, and `local( ... )`.
+   `{ ... }`, `( ... )`, `local( ... )`, `I()`, and `identity()`.
 
 
 # Version 0.0.4 (2025-09-26)