+++ NEWS.md
@@ -1,4 +1,4 @@
-# Version 0.0.4-9000 (2025-10-07)
+# Version 0.0.5 (2025-11-26)
 
  * Add support for futurizing calls nested in, one or more layers, of
    `{ ... }`, `( ... )`, `local( ... )`, `I()`, and `identity()`, e.g.